+++ Client/Source/GUI/DropDownList.h
@@ -74,5 +74,5 @@ public:
 	/*! Removes all normal stuff from the renderer*/
 	virtual bool RemoveFromRenderer(GraphicsEngine* ge);
 
-	void Resize(int oldWindowWidth, int oldWindowHeight, int windowWidth, int windowHeight);
+	void Resize(float oldWindowWidth, float oldWindowHeight, float windowWidth, float windowHeight);
 };
